requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.9
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
python-decouple==3.8
python-dotenv==1.2.1
gunicorn==22.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
sib-api-v3-sdk==7.6.0
